[WFCORE-6206] Add remove-custom-patch operation and remove this functionality from the clean operation
