+++ installation-manager/src/main/java/org/wildfly/core/instmgr/cli/CustomPatchRemoveCommand.java
@@ -1,0 +1,48 @@
+package org.wildfly.core.instmgr.cli;
+
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP;
+
+import org.aesh.command.CommandDefinition;
+import org.aesh.command.CommandException;
+import org.aesh.command.CommandResult;
+import org.aesh.command.option.Option;
+import org.jboss.as.cli.CommandContext;
+import org.jboss.as.controller.client.ModelControllerClient;
+import org.jboss.as.controller.client.Operation;
+import org.jboss.as.controller.client.OperationBuilder;
+import org.jboss.dmr.ModelNode;
+import org.wildfly.core.cli.command.aesh.CLICommandInvocation;
+import org.wildfly.core.instmgr.InstMgrConstants;
+import org.wildfly.core.instmgr.InstMgrCustomPatchRemoveHandler;
+
+@CommandDefinition(name = "remove-custom-patch", description = "Removes a custom patch and its channel from the server.")
+public class CustomPatchRemoveCommand extends AbstractInstMgrCommand {
+
+    @Option(name = "manifest", required = true)
+    private String manifestGA;
+
+    @Override
+    protected Operation buildOperation() throws CommandException {
+        final ModelNode op = new ModelNode();
+        final OperationBuilder operationBuilder = OperationBuilder.create(op);
+
+        op.get(OP).set(InstMgrCustomPatchRemoveHandler.DEFINITION.getName());
+        op.get(InstMgrConstants.MANIFEST).set(manifestGA);
+
+        return operationBuilder.build();
+    }
+
+    @Override
+    public CommandResult execute(CLICommandInvocation commandInvocation) throws CommandException, InterruptedException {
+        final CommandContext ctx = commandInvocation.getCommandContext();
+        final ModelControllerClient client = ctx.getModelControllerClient();
+        if (client == null) {
+            ctx.printLine("You are disconnected at the moment. Type 'connect' to connect to the server or 'help' for the list of supported commands.");
+            return CommandResult.FAILURE;
+        }
+
+        this.executeOp(ctx, this.host);
+        ctx.printLine("Custom Patch successfully removed");
+        return CommandResult.SUCCESS;
+    }
+}
