+++ installation-manager/src/main/resources/org/wildfly/core/instmgr/LocalDescriptions.properties
@@ -47,7 +47,6 @@ installation-manager.prepare-revert.maven-repo-files=The location of the local M
 # clean Operation
 installation-manager.clean=Clean installation manager content. If no arguments are provided, executing this operation will delete any existing candidate server along with their associated content.
 installation-manager.clean.work-dir=Deletes a specific installation manager work directory.
-installation-manager.clean.custom-patch-manifest=Manifest GA (GroupId:ArtifactId) of the custom patch to be deleted. When using this argument, this operation deletes the custom patch defined by this manifest coordinates which was uploaded to the server and unsubscribe this installation from the channel associated to this custom patch.
 
 # Clone Export
 installation-manager.clone-export=Export the installation metadata.
@@ -67,6 +66,8 @@ installation-manager.channel-remove.name=Name of the channel to be removed.
 installation-manager.custom-patch.upload-custom-patch=Uploads a custom patch zip file and subscribes the current installation to a channel to consume it.
 installation-manager.custom-patch.upload-custom-patch.manifest=Channel Manifest maven coordinates associated to the custom patch. Expected format is GA (GroupId:ArtifactId)
 installation-manager.custom-patch.upload-custom-patch.custom-patch-file=custom patch file
-installation-manager.custom-patch.manifest.gav=Maven manifest coordinates following the format GroupId:ArtifactId:Version. GroupId and ArtifactId elements are mandatory. Version value is optional to stick to a given manifest version (instead of requiring the latest version of that manifest). In the absence of this version, the latest version of the manifest will be determined based on the Maven repository metadata.
-installation-manager.custom-patch.manifest.url=URL where the manifest file can be found.
+
+# Remove Custom Patch
+installation-manager.custom-patch.remove-custom-patch=Uploads a custom patch zip file and subscribes the current installation to a channel to consume it.
+installation-manager.custom-patch.remove-custom-patch.manifest=Channel Manifest maven coordinates associated to the custom patch. Expected format is GA (GroupId:ArtifactId)
