+++ installation-manager/src/main/resources/org/wildfly/core/instmgr/cli/command_resources.properties
@@ -23,7 +23,7 @@ installer.description=\
 
 # GENERIC HOST
 installer.abstract-inst-mgr-cmd.option.host.description=\
-  Name of the host where the channel will be added. This option is required only in the domain mode and is not available in the standalone mode.
+  Name of the host where this operation will be executed. This option is required only in the domain mode and is not available in the standalone mode.
 
 
 # CHANNEL ADD
@@ -75,10 +75,6 @@ installer.channel-remove.option.channel-name.description=\
 installer.clean.description=\
   Removes all the custom patches, temporary files, and prepared installations that are ready to be applied. When you run this command without any options, all \
   the prepared and pending installations that are yet to be applied to the server are deleted.
-
-installer.clean.option.custom-patch-manifest.description=\
-  Manifest artifact location expressed as GA (GroupId:aritfactId) maven coordinate.
-
 
 # HISTORY
 installer.history.description=\
@@ -178,7 +174,7 @@ installer.update.option.headers.description=\
   the server candidate. You can use the operation headers to increase the default blocking timeout which is 300 seconds.
 
 
-# CUSTOM PATCHES
+# UPLOAD CUSTOM PATCHES
 installer.upload-custom-patch.description=\
   Uploads a custom patch to the base server. It also creates a new channel that will provide this custom patch and subscribes the installation to this channel. \
   After uploading a custom patch if you want to remove it, you can use the clean operation to remove it from the server.
@@ -187,5 +183,14 @@ installer.upload-custom-patch.option.custom-patch-file.description=\
   Location of the custom patch file.
 
 installer.upload-custom-patch.option.manifest.description=\
-  Location of the channel manifest artifact. A manifest defines the versions and Artifacts that will be available from this channel. Specify the location as a file \
+  Location of the channel manifest artifact associated with the custom patch that will be uploaded. A manifest defines the versions and Artifacts that will be available from the channel that consumes this custom patch. Specify the location as a file \
   path, URL or Maven GAV coordinate (groupId:artifactId:version). If the manifest version is not mentioned in the GAV coordinates, the latest available manifest is used.
+
+
+# REMOVE CUSTOM PATCHES
+installer.remove-custom-patch.description=\
+  Removes a custom patch from the base server. It also removes the channel that provides this custom patch and unsubscribes the installation from this channel.
+
+installer.remove-custom-patch.option.manifest.description=\
+  Location of the channel manifest artifact associated with custom patch that will be removed. A manifest defines the versions and Artifacts that will be available from the channel that consumes this custom patch. \
+  Specify the location as a Maven GA coordinate (groupId:artifactId).
